requirements.txt
# --------- pytorch --------- #
torch>=1.13.0
torchvision>=0.14.0
pytorch-lightning>=1.8.3
torchmetrics>=0.11.0

# --------- hydra --------- #
hydra-core==1.3.1
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
hydra-zen>=0.9.0

#---------- others -----------#
scipy>=1.10.0
matplotlib>=3.5.2
scikit-learn>=1.0.2
scikit-image>=0.19.2
seaborn>=0.11.2
jupyter>=1.0.0
h5py>=3.6.0
pyrootutils     # standardizing the project root setup
# pre-commit    # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
